+++ math/include/Point3.h
@@ -9,10 +9,10 @@ class Vector3;
 
 class Point3{
 public:
-    double x, y, z;
+    float x, y, z;
     Point3(): x(0.0f), y(0.0f), z(0.0f) {}
-    Point3(double x, double y, double z): x(x), y(y), z(z) {};
-    Point3(std::array<double,3> &a) : x(a[0]), y(a[1]), z(a[2]) {};
+    Point3(float x, float y, float z): x(x), y(y), z(z) {};
+    Point3(std::array<float,3> &a) : x(a[0]), y(a[1]), z(a[2]) {};
     Point3(Vector3&& v);
 
     Point3 operator+(const Point3& other) const {
@@ -27,15 +27,15 @@ public:
         return Point3(x * other.x, y * other.y, z * other.z);
     }
 
-    Point3 operator*(double scalar) const {
+    Point3 operator*(float scalar) const {
         return Point3(x * scalar, y * scalar, z * scalar);
     }
 
-    Point3 operator/(double scalar) const {
+    Point3 operator/(float scalar) const {
         return Point3(x / scalar, y / scalar, z / scalar);
     }
 
-    double operator[](int i) const {
+    float operator[](int i) const {
         switch (i) {
             case 0:
                 return x;
